requirements.txt
Django==5.2.5
djangorestframework==3.15.2
drf-yasg==1.21.7
django-cors-headers==4.4.0
PyYAML==6.0.1
inflection==0.5.1
packaging==24.1
ruamel.yaml==0.18.6
setuptools==70.0.0
uritemplate==4.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
whitenoise==6.5.0
dj-database-url==2.1.0
